// src/api/model/consulta.ts
/**
 * Plenos API
 * API para gestionar consultas de usuarios.
 *
 * OpenAPI spec version: 3.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Consulta { 
    idPleno: number;
    idConsulta: number;
}